requirements.txt
# Love-Unlimited - Memory Sovereignty Hub
# Python dependencies

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebRTC
aiortc==1.6.0

# TTS
pyttsx3==2.90

# Memory Storage
chromadb==0.4.22
sentence-transformers==2.3.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Data Models
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Short-term memory (optional Redis support)
# redis==5.0.1
# aioredis==2.0.1

# Utilities
httpx==0.26.0
aiohttp==3.9.1
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
